+++ src/Components/Activities/ActivitiesForm.js
@@ -1,17 +1,73 @@
+import React, { useState, useEffect } from "react";
 import { CKEditor } from "@ckeditor/ckeditor5-react";
 import ClassicEditor from "@ckeditor/ckeditor5-build-classic";
+import { useParams } from "react-router-dom";
 import "../FormStyles.css";
+import axios from "axios";
 
-const ActivitiesForm = ({
-  activity,
-  handleChangeName,
-  handleChangeImage,
-  handleChangeDescription,
-  handleSubmit,
-}) => {
+const toDataURL = (blob) =>
+  new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.onloadend = () => resolve(reader.result);
+    reader.onerror = reject;
+    reader.readAsDataURL(blob);
+  });
+
+const ActivitiesForm = () => {
+  const { id } = useParams();
+  const [activity, setActivity] = useState({});
   const { name, description, image } = activity;
+
+  useEffect(() => {
+    if (id) {
+      axios.get(`http://ongapi.alkemy.org/api/activities/${id}`).then((res) => {
+        setActivity(res.data.data);
+      });
+    }
+  }, [id]);
+
+  const handleChangeName = (e) => {
+    setActivity((prevActivity) => ({ ...prevActivity, name: e.target.value }));
+  };
+
+  const handleChangeImage = (e) => {
+    const img = URL.createObjectURL(e.target.files[0]);
+    axios
+      .get(img, { responseType: "blob" })
+      .then((response) => toDataURL(response.data))
+      .then((encoded) => {
+        setActivity((prevActivity) => ({ ...prevActivity, image: encoded }));
+      });
+  };
+
+  const handleChangeDescription = (e, editor) => {
+    const data = editor.getData();
+    setActivity((prevActivity) => ({ ...prevActivity, description: data }));
+  };
+
+  const handleSubmitCreate = (e) => {
+    e.preventDefault();
+    axios.post("http://ongapi.alkemy.org/api/activities", {
+      name,
+      description,
+      image,
+    }); //TODO: Controlar errores (Catch)
+  };
+
+  const handleSubmitUpdate = (e) => {
+    e.preventDefault();
+    axios.put(`http://ongapi.alkemy.org/api/activities/${id}`, {
+      name,
+      description,
+      image,
+    }); //TODO: Controlar errores (Catch)
+  };
+
   return (
-    <form className="form-container" onSubmit={handleSubmit}>
+    <form
+      className="form-container"
+      onSubmit={id ? handleSubmitUpdate : handleSubmitCreate}
+    >
       <input
         className="input-field"
         type="text"
